[update] cgi_parseをnamespace httpに移動しました、変更量が多いので以下に詳細を書きます
- cgi_parseをnamespace httpに移動
- cgi_requestを新規作成し、今までの定数やCgiRequestの定義を中に移動
