+++ srcs/cgi/cgi.cpp
@@ -1,5 +1,5 @@
 #include "cgi.hpp"
-#include "cgi_parse.hpp"
+#include "cgi_request.hpp"
 #include "http_exception.hpp"
 #include "http_message.hpp"
 #include "status_code.hpp"
