+++ srcs/cgi/cgi_request.cpp
@@ -1,0 +1,23 @@
+#include "cgi_request.hpp"
+
+namespace cgi {
+
+const std::string AUTH_TYPE         = "AUTH_TYPE";
+const std::string CONTENT_LENGTH    = "CONTENT_LENGTH";
+const std::string CONTENT_TYPE      = "CONTENT_TYPE";
+const std::string GATEWAY_INTERFACE = "GATEWAY_INTERFACE";
+const std::string PATH_INFO         = "PATH_INFO";
+const std::string PATH_TRANSLATED   = "PATH_TRANSLATED";
+const std::string QUERY_STRING      = "QUERY_STRING";
+const std::string REMOTE_ADDR       = "REMOTE_ADDR";
+const std::string REMOTE_HOST       = "REMOTE_HOST";
+const std::string REMOTE_IDENT      = "REMOTE_IDENT";
+const std::string REMOTE_USER       = "REMOTE_USER";
+const std::string REQUEST_METHOD    = "REQUEST_METHOD";
+const std::string SCRIPT_NAME       = "SCRIPT_NAME";
+const std::string SERVER_NAME       = "SERVER_NAME";
+const std::string SERVER_PORT       = "SERVER_PORT";
+const std::string SERVER_PROTOCOL   = "SERVER_PROTOCOL";
+const std::string SERVER_SOFTWARE   = "SERVER_SOFTWARE";
+
+} // namespace cgi
